login system, some bug fixes and clock widget

+++ server/headers/statstype.hpp
@@ -2,4 +2,4 @@
 
 #include <cstdint>
 
-typedef uint16_t stattype;
+typedef int stattype;
